==> P1_snfpy.ipynb <==
from snf import datasets
import numpy as np
import pandas as pd
import snf
from snf import compute
import sklearn

# digits = datasets.load_digits()
# digits.keys()

# for arr in digits.data:
#     print(arr.shape)

# groups, samples = np.unique(digits.labels, return_counts=True)

# for grp, count in zip(groups, samples):
#     print('Group {:.0f}: {} samples'.format(grp, count))

########## Drug Data
# Number of Drug = 708

##Similarity
Drugs_chemStracture = pd.read_csv('data/Drug disease/drug sim/Similarity_Matrix_Drugs chem stracture.txt',
                                     sep=' ', header=None)
# simDrugs_chemStracture = simDrugs_chemStracture.dropna(axis=1)
Drugs_chemStracture.replace(np.nan, 0, inplace=True)
print(Drugs_chemStracture.head(), Drugs_chemStracture.shape)


# Drug features
sideeffect = pd.read_csv('data/Drug disease/drug sim/mat_drug_side effect.txt',sep=' ', header=None)
sideeffect = sideeffect.replace(np.nan,0)
print(sideeffect.head(), sideeffect.shape)

protein = pd.read_csv('data/Drug disease/drug sim/mat_drug_protein.txt',sep=' ', header=None)
protein.replace(np.nan, 0, inplace=True)

print(protein.head(), protein.shape)


########## Disease Data
# Number of Disease = 5603
# Disease Feature

diseasePhenotype = pd.read_csv('data/Drug disease/disease sim/mat_disease Phenotype.txt',sep='\t', header=None)
diseasePhenotype.replace(np.nan, 0, inplace=True)
print(diseasePhenotype.head(), diseasePhenotype.shape)


disease_protein = pd.read_csv('data/Drug disease/disease sim/mat_disease_protein.txt',sep='\t', header=None)
disease_protein.replace(np.nan, 0, inplace=True)
print(disease_protein.head(), disease_protein.shape)

# Phenotype = pd.read_csv('data/Drug disease/disease sim/mat_disease Phenotype.txt',sep='\t', header=None)
# Phenotype = Phenotype.drop([1627],axis=1)
# print(Phenotype.head())
# protein = pd.read_csv('data/Drug disease/disease sim/mat_disease_protein.txt',sep='\t', header=None)
# print(protein.head())

'''Similarity calculation based on feature data, 
and then, returning two ndarray similarities for drugs and diseases'''

# print(offsidesF.isnull().sum().sum())
protein = np.array(protein.values, dtype=np.float64)
Drugs_chemStracture = np.array(Drugs_chemStracture.values, dtype=np.float64)
sideeffect = np.array(sideeffect.values, dtype=np.float64)

diseasePhenotype = np.array(diseasePhenotype.values, dtype=np.float64)
disease_protein = np.array(disease_protein.values, dtype=np.float64)

proteinSim = sklearn.metrics.pairwise.cosine_similarity(protein, Y=None, dense_output=True)
Drugs_chemStracture = sklearn.metrics.pairwise.cosine_similarity(Drugs_chemStracture, Y=None, dense_output=True)
# chemstractureSim = chemstracture.copy()
sideeffectSim = sklearn.metrics.pairwise.cosine_similarity(protein, Y=None, dense_output=True)


disease_proteinSim = sklearn.metrics.pairwise.cosine_similarity(disease_protein, Y=None, dense_output=True)
diseasePhenotypeSim = sklearn.metrics.pairwise.cosine_similarity(diseasePhenotype, Y=None, dense_output=True)

diseaseSimData = [disease_proteinSim, diseasePhenotypeSim]
drugSimData = [Drugs_chemStracture, proteinSim, sideeffectSim]
del Drugs_chemStracture, proteinSim, sideeffectSim, disease_proteinSim, diseasePhenotypeSim 
del disease_protein, diseasePhenotype, protein, sideeffect
# drugSimData
# diseaseSimData 

# affinity_networks = compute.make_affinity(data, metric='cosine', K=9150, mu=0.5,normalize=True)
# # simcosine = pd.DataFrame(affinity_networks)
# # simcosine.to_csv("cosineSim.csv",index=False)

# len(affinity_networks[1])

# for grp, count in zip(groups, samples):
#     print('Group {:.0f}: {} samples'.format(grp, count))# chem =  np.array(pd.read_csv('chem_Jacarrd_sim.csv', header=-1))
# enzyme =  np.array(pd.read_csv('enzyme_Jacarrd_sim.csv', header=-1))
# # chem = array(chem,enzyme)
# affinity_network = [chem,enzyme]

fused_drugSim = snf.snf(drugSimData, K=20)
fused_diseaseSim = snf.snf(diseaseSimData, K=20)

print(fused_drugSim.shape, fused_diseaseSim.shape)

# fused_diseaseSim

# fused_drugSim

# digits.data

pd.DataFrame(fused_drugSim).to_csv(
    'data/Drug disease/Drug_CosineSNF(chemStracture_DrugProtein_sideEffect).csv', 
                                      index = False,header=False)

pd.DataFrame(fused_diseaseSim).to_csv(
    'data/Drug disease/Disease_CosineSNF(DisPhenotype_DisProteinm).csv', 
                                      index = False,header=False)
