==> P7-7_choosingProbableZeros.ipynb <==
import pandas as pd
import numpy as np
import random

def reduceList( DrugPairs):
#     print(dPairs.head())
    i = 0
    newList = []
    seenList = []
    while i < len(DrugPairs.iloc[:,0]):
        d1, d2 = DrugPairs.iloc[i][1], DrugPairs.iloc[i][2]
        buf = DrugPairs[DrugPairs[1]==d2][DrugPairs[2]==d1]
#         print(buf)
        if len(buf) == 1 and (i not in seenList) and (list(buf.index)[0] not in seenList):
            newList.append(i)
            newList.append(list(buf.index)[0])
            seenList.append(list(buf.index)[0])
            seenList.append(i)
        else:
            seenList.append(i)
        i += 1
    return newList

data = pd.read_csv('triple_cosineSNF(zeros)_rivised.csv')

predict = pd.read_csv('predict_(-1 and +1 model)all_rivised.csv')
predict.head()

zeroList = list(predict[predict['0']<0.2][predict['1']<0.2].index)
len(zeroList)

probZeroDrug = data.iloc[zeroList,:]
probZeroDrug.head()
print(probZeroDrug.shape)

probZeroDrug.iloc[:,0:3].to_csv('probZeroDrugpairs_rivised.csv',header = False)
drugsPairs = pd.read_csv('probZeroDrugpairs_rivised.csv',header=None)
r, c = drugsPairs.shape
print(r, c)
drugsPairs.head()

newList = reduceList(drugsPairs.copy())
# print(newDegList)
len(newList)

probZeroDrug.iloc[newList,:].to_csv('probZeroDrugBothPair_.2_.2_revised.csv',index=False)

probZeroDrugBothPair = pd.read_csv('probZeroDrugBothPair_.2_.2_revised.csv')

drugsPairs = probZeroDrugBothPair.iloc[:,0:3]
r, c = drugsPairs.shape
print(r, c)
drugsPairs.head()

indexes = []
bufpairs = []
for i in range(int(10000)):
    candInd = np.random.randint(r/2)

    while (int(2 * candInd) in indexes):
        candInd = np.random.randint(r/2)
    indexes.append(int(2 * candInd))
    indexes.append(int(2 * candInd + 1))

print(len(indexes))


drugsPairs.iloc[indexes,:]

# index = drugsPairs.values[:,0]
# random.shuffle(index)
# drugsPairsShuffled =  drugsPairs.iloc[index]

checked = False
problems = []
i = 0
while(i +1 < len(indexes)):

    if (drugsPairs.iloc[indexes[i],0]==drugsPairs.iloc[indexes[i+1],1])and(drugsPairs.iloc[indexes[i],1]==drugsPairs.iloc[indexes[i+1],0]):
        i += 2

    else:        
        problems.append([i,drugsPairs.iloc[i,:]])

        del indexes[i]

len(problems)

probZeroDrug.iloc[indexes,:].to_csv('probZeroDrug20000_.2_.2_revised.csv',index=False)

probZeroDrug.iloc[indexes,:].iloc[:18000,:].to_csv('probZeroDrugTrain_revised.csv',index=False)

probZeroDrug.iloc[indexes,:].iloc[18000:,:].to_csv('probZeroDrugTest_revised.csv',index=False)
